# Cut down the LUT size by 4 KB