--- a/scripts/sse_utf8_utf16_decode.py
+++ b/scripts/sse_utf8_utf16_decode.py
@@ -66,6 +66,7 @@ def buildshuf12_twobytes(sizes):
             answer[2*i] = pos + 1
             answer[2*i+1] = pos
             pos += 2
+    answer[15] = 0xF0 | sum(sizes)
     return answer
 
 def buildshuf123_threebytes(sizes):
@@ -90,6 +91,7 @@ def buildshuf123_threebytes(sizes):
             answer[4*i+2] = pos
             answer[4*i+3] = 0xff
             pos += 3
+    answer[15] = 0xF0 | sum(sizes)
     return answer
 
 def buildshuf1234_fourbytes(sizes):
@@ -120,6 +122,7 @@ def buildshuf1234_fourbytes(sizes):
             answer[4*i+2] = pos + 1
             answer[4*i+3] = pos
             pos += 4
+    answer[15] = 0xF0 | sum(sizes)
     return answer
 
 
@@ -163,23 +166,20 @@ def main():
     if(easy_case12(sizes)):
         z1 = grab_easy_case12_code_point_size(sizes)
         idx = index[tuple(z1)]
-        s = sum(z1)
-        arrg.append((idx,s))
+        arrg.append(idx)
     elif(easy_case123(sizes)):
         z1 = grab_easy_case123_code_point_size(sizes)
         idx = index[tuple(z1)]
-        s = sum(z1)
-        arrg.append((idx,s))
+        arrg.append(idx)
     elif(easy_case1234(sizes)):
         z1 = grab_easy_case1234_code_point_size(sizes)
         idx = index[tuple(z1)]
-        s = sum(z1)
-        arrg.append((idx,s))
+        arrg.append(idx)
     else:
         # we are in error, use a bogus index
-        arrg.append((209,12))
-  print("const uint8_t utf8bigindex["+str(len(arrg))+"][2] = ")
-  print(cpp_arrayarray_initializer(arrg), end=";\n")
+        arrg.append(209)
+  print("const uint8_t utf8bigindex["+str(len(arrg))+"] = ")
+  print(cpp_array_initializer(arrg), end=";\n")
 
 
 if __name__ == '__main__':