--- a/scripts/sse_utf8_utf16_decode.py
+++ b/scripts/sse_utf8_utf16_decode.py
@@ -66,7 +66,6 @@ def buildshuf12_twobytes(sizes):
             answer[2*i] = pos + 1
             answer[2*i+1] = pos
             pos += 2
-    answer[15] = 0xF0 | sum(sizes)
     return answer
 
 def buildshuf123_threebytes(sizes):
@@ -91,7 +90,6 @@ def buildshuf123_threebytes(sizes):
             answer[4*i+2] = pos
             answer[4*i+3] = 0xff
             pos += 3
-    answer[15] = 0xF0 | sum(sizes)
     return answer
 
 def buildshuf1234_fourbytes(sizes):
@@ -122,9 +120,8 @@ def buildshuf1234_fourbytes(sizes):
             answer[4*i+2] = pos + 1
             answer[4*i+3] = pos
             pos += 4
-    answer[15] = 0xF0 | sum(sizes)
-    return answer
 
+    return answer
 
 def main():
   easycase12 = set()
@@ -161,26 +158,35 @@ def main():
       index[t] = c
       c = c + 1
   arrg=[]
+  sizes_lut = []
+
   for x in range(1<<12):
     sizes = compute_code_point_size(x)
     if(easy_case12(sizes)):
         z1 = grab_easy_case12_code_point_size(sizes)
         idx = index[tuple(z1)]
         arrg.append(idx)
+        sizes_lut.append(sum(z1))
     elif(easy_case123(sizes)):
         z1 = grab_easy_case123_code_point_size(sizes)
         idx = index[tuple(z1)]
         arrg.append(idx)
+        sizes_lut.append(sum(z1))
     elif(easy_case1234(sizes)):
         z1 = grab_easy_case1234_code_point_size(sizes)
         idx = index[tuple(z1)]
         arrg.append(idx)
+        sizes_lut.append(sum(z1))
     else:
         # we are in error, use a bogus index
         arrg.append(209)
+        sizes_lut.append(12)
   print("const uint8_t utf8bigindex["+str(len(arrg))+"] = ")
   print(cpp_array_initializer(arrg), end=";\n")
 
+  print("const uint8_t utf8lenindex["+str(len(sizes_lut) // 2) + "] = ")
+  print(cpp_array_initializer(sizes_lut[:len(sizes_lut)//2]), end=";\n")
 
 if __name__ == '__main__':
     main()
+