--- a/src/haswell/avx2_convert_utf8_to_utf16.cpp
+++ b/src/haswell/avx2_convert_utf8_to_utf16.cpp
@@ -71,9 +71,10 @@ size_t convert_masked_utf8_to_utf16(const char *input,
   }
 
   const uint8_t idx =
-      simdutf::tables::utf8_to_utf16::utf8bigindex[input_utf8_end_of_code_point_mask][0];
+      simdutf::tables::utf8_to_utf16::utf8bigindex[input_utf8_end_of_code_point_mask];
   const uint8_t consumed =
-      simdutf::tables::utf8_to_utf16::utf8bigindex[input_utf8_end_of_code_point_mask][1];
+      simdutf::tables::utf8_to_utf16::shufutf8[idx][15] & 0xF;
+
   if (idx < 64) {
     // SIX (6) input code-words
     // this is a relatively easy scenario