--- a/src/westmere/sse_convert_utf8_to_utf32.cpp
+++ b/src/westmere/sse_convert_utf8_to_utf32.cpp
@@ -21,6 +21,8 @@ size_t convert_masked_utf8_to_utf32(const char *input,
   const __m128i in = _mm_loadu_si128((__m128i *)input);
   const uint16_t input_utf8_end_of_code_point_mask =
       utf8_end_of_code_point_mask & 0xfff;
+  const uint16_t input_utf8_len_mask =
+      utf8_end_of_code_point_mask & 0x7ff;
   if(((utf8_end_of_code_point_mask & 0xffff) == 0xffff)) {
     // We process the data in chunks of 16 bytes.
     _mm_storeu_si128(reinterpret_cast<__m128i *>(utf32_output), _mm_cvtepu8_epi32(in));
@@ -66,8 +68,8 @@ size_t convert_masked_utf8_to_utf32(const char *input,
 
   const uint8_t idx =
       tables::utf8_to_utf16::utf8bigindex[input_utf8_end_of_code_point_mask];
-  const uint8_t consumed =
-      tables::utf8_to_utf16::shufutf8[idx][15] & 0xF;
+  uint8_t consumed =
+      tables::utf8_to_utf16::utf8lenindex[input_utf8_len_mask];
 
   if (idx < 64) {
     // SIX (6) input code-words
@@ -84,6 +86,8 @@ size_t convert_masked_utf8_to_utf32(const char *input,
     _mm_storeu_si128(reinterpret_cast<__m128i *>(utf32_output), _mm_cvtepu16_epi32(composed));
     _mm_storeu_si128(reinterpret_cast<__m128i *>(utf32_output+4), _mm_cvtepu16_epi32(_mm_srli_si128(composed,8)));
     utf32_output += 6; // We wrote 12 bytes, 6 code points.
+    // Corner case in half LUT
+    if (input_utf8_end_of_code_point_mask == 0xaaa) consumed = 12;
   } else if (idx < 145) {
     // FOUR (4) input code-words
     const __m128i sh =