This encodes the length in the last element of shufutf8, similar to how it is in the UTF-16 table, but it is further minimized by making it 0xF0 | length since the last byte will always be zero.