@lemire 2023-05-03
This PR increases the length of the data dependency chain, so it is at risk of leading to lower instructions per cycle performance.

Let us try it out. Build the library and benchmarks into the build subdirectory and grab the data files from https://github.com/lemire/unicode_lipsum (putting unicode_lipsum directly as a subdirectory).


Current main branch:

```
./build/benchmarks/benchmark -P convert_utf8_to_utf16+westmere -F unicode_lipsum/lipsum/*.utf8.txt | grep GB
   2.252 GB/s (3.6 %)    1.262 Gc/s     1.78 byte/char
   3.196 GB/s (2.3 %)    1.074 Gc/s     2.98 byte/char
   1.613 GB/s (2.6 %)    0.403 Gc/s     4.00 byte/char
   2.249 GB/s (3.0 %)    1.262 Gc/s     1.78 byte/char
   1.914 GB/s (4.1 %)    0.713 Gc/s     2.69 byte/char
   3.095 GB/s (3.6 %)    1.067 Gc/s     2.90 byte/char
   2.130 GB/s (3.2 %)    0.868 Gc/s     2.45 byte/char
  21.916 GB/s (6.8 %)   21.916 Gc/s     1.00 byte/char
   2.261 GB/s (4.5 %)    1.251 Gc/s     1.81 byte/char
```

Your PR:

```
$ ./build/benchmarks/benchmark -P convert_utf8_to_utf16+westmere -F unicode_lipsum/lipsum/*.utf8.txt |grep GB
   1.834 GB/s (3.8 %)    1.028 Gc/s     1.78 byte/char
   2.995 GB/s (4.7 %)    1.006 Gc/s     2.98 byte/char
   1.617 GB/s (3.2 %)    0.404 Gc/s     4.00 byte/char
   1.828 GB/s (2.7 %)    1.025 Gc/s     1.78 byte/char
   1.764 GB/s (4.6 %)    0.657 Gc/s     2.69 byte/char
   2.814 GB/s (3.9 %)    0.970 Gc/s     2.90 byte/char
   1.762 GB/s (4.1 %)    0.718 Gc/s     2.45 byte/char
  22.801 GB/s (6.5 %)   22.801 Gc/s     1.00 byte/char
   1.885 GB/s (2.8 %)    1.043 Gc/s     1.81 byte/char
```

As you'd expect, the performance does not always strike. From first principles, you'd expect it to mostly appear in "UTF-8 rich" inputs (where the size of the read offset is hard to predict). And indeed, that is what we see in these results: there is a 20% performance regression in several cases.

We do want to have small tables, but ideally, it should not create longer dependency chains. The "read offset" is on a critical path. We want to have it as soon as possible, upfront, and not in the latter part of the process. E.g., it should be immediately available from the mask, and not require an intermediate access.