@easyaspi314 2023-05-03
Hmm. Well I did some tests and apparently with the exception of `mask == 0x888` (which lands in the rarer 4 byte path) the first and second halves of the size LUT are the same.

So the table can be cut in half with one `and` and a simple branch on the slow path.

For a more extreme solution packing nibbles is an option but that has extra branches.