@lemire 2023-05-03
@easyaspi314 Sure!!! Saving 2 kB if it is performance neutral would be huge. Can you try it out ?

I recommend you run benchmarks too.