@easyaspi314 2023-05-03
Well an initial test on an ARM Cortex-X1 (yes it is my phone) + Clang 16 shows about 5% overhead and probably can be tuned to be less.

I can get you x86 benchmarks once I clean it up.

Also, perhaps we could have a build option that either uses the last byte in the shuffle mask or the dedicated LUT, since there is no harm in leaving the length in the table. People who compile with -Os would much rather take 2 KB over 10-20% performance.