@clausecker 2023-05-03
These 5% are more than compensated for by reducing cache pressure.  Remember, in real world applications, transcoding strings is only a small part of what an application does.  Not trashing 4 kB of L1 cache is definitely worth a 5% perf loss in my opinion.