@lemire 2023-05-04
> is definitely worth a 5% perf loss

Glancing at the code, I am not sure that this should cause a 5% perf loss. It is fairly difficult to measure a 5% difference, as we all know.