@easyaspi314 2023-05-04
Cortex-X1, original, clang 16.0.2
```
   1.780 GB/s (3.2 %)    0.997 Gc/s     1.78 byte/char 
   2.899 GB/s (2.2 %)    0.974 Gc/s     2.98 byte/char 
   1.226 GB/s (4.7 %)    0.306 Gc/s     4.00 byte/char 
   1.774 GB/s (6.5 %)    0.995 Gc/s     1.78 byte/char 
   1.665 GB/s (3.2 %)    0.620 Gc/s     2.69 byte/char 
   2.521 GB/s (3.5 %)    0.869 Gc/s     2.90 byte/char 
   1.627 GB/s (2.8 %)    0.663 Gc/s     2.45 byte/char 
  24.559 GB/s (19.4 %)   24.559 Gc/s     1.00 byte/char - jittery af
   1.796 GB/s (5.2 %)    0.994 Gc/s     1.81 byte/char 
```

Half LUT: There seem to be some gains and losses but this can also be assumed to jitter. As expected, the path with the branch (which I was wrong, it is the 2 byte path) is a tiny bit slower.
```
   1.635 GB/s (2.5 %)    0.916 Gc/s     1.78 byte/char -> -8%
   3.333 GB/s (4.6 %)    1.120 Gc/s     2.98 byte/char -> +13%
   1.314 GB/s (5.0 %)    0.328 Gc/s     4.00 byte/char -> +6%
   1.620 GB/s (2.9 %)    0.909 Gc/s     1.78 byte/char -> -9%
   1.711 GB/s (2.4 %)    0.637 Gc/s     2.69 byte/char -> +3%
   2.654 GB/s (1.1 %)    0.915 Gc/s     2.90 byte/char -> +5%
   1.581 GB/s (1.8 %)    0.645 Gc/s     2.45 byte/char -> -5%
  23.485 GB/s (10.0 %)   23.485 Gc/s     1.00 byte/char -> jittery af
   1.635 GB/s (1.1 %)    0.905 Gc/s     1.81 byte/char -> -6%
```
Will edit with SSE4 and AVX2 benches soon