@lemire 2023-05-04
Here are my results (compare with above) on this PR....


Before...

```
./build/benchmarks/benchmark -P convert_utf8_to_utf16+westmere -F unicode_lipsum/lipsum/*.utf8.txt | grep GB
   2.252 GB/s (3.6 %)    1.262 Gc/s     1.78 byte/char
   3.196 GB/s (2.3 %)    1.074 Gc/s     2.98 byte/char
   1.613 GB/s (2.6 %)    0.403 Gc/s     4.00 byte/char
   2.249 GB/s (3.0 %)    1.262 Gc/s     1.78 byte/char
   1.914 GB/s (4.1 %)    0.713 Gc/s     2.69 byte/char
   3.095 GB/s (3.6 %)    1.067 Gc/s     2.90 byte/char
   2.130 GB/s (3.2 %)    0.868 Gc/s     2.45 byte/char
  21.916 GB/s (6.8 %)   21.916 Gc/s     1.00 byte/char
   2.261 GB/s (4.5 %)    1.251 Gc/s     1.81 byte/char
```

After:

```
$ ./build/benchmarks/benchmark -P convert_utf8_to_utf16+westmere -F unicode_lipsum/lipsum/*.utf8.txt |grep GB
   2.073 GB/s (3.4 %)    1.162 Gc/s     1.78 byte/char
   2.969 GB/s (5.5 %)    0.997 Gc/s     2.98 byte/char
   1.585 GB/s (3.2 %)    0.396 Gc/s     4.00 byte/char
   2.066 GB/s (3.3 %)    1.159 Gc/s     1.78 byte/char
   1.873 GB/s (3.0 %)    0.698 Gc/s     2.69 byte/char
   2.948 GB/s (3.7 %)    1.016 Gc/s     2.90 byte/char
   2.139 GB/s (2.7 %)    0.872 Gc/s     2.45 byte/char
  22.419 GB/s (5.4 %)   22.419 Gc/s     1.00 byte/char
   2.071 GB/s (2.3 %)    1.146 Gc/s     1.81 byte/char
```

Binary size, before...

```
$ ls -al build/src/*.a
-rw-r--r--  1 lemire  staff  382216  3 May 12:12 build/src/libsimdutf.a
```

After...

```
$ ls -al build/src/*.a
-rw-r--r--  1 lemire  staff  381736  3 May 21:55 build/src/libsimdutf.a
```

So this PR appears to shave 480 bytes from the library on my machine right now. (So 0.1%.)