@easyaspi314 2023-05-04
> So this PR appears to shave 480 bytes from the library on my machine right now. (So 0.1%.)

The `.a` file isn't really indicative of the actual binary size, it is recommended is to do `size simdutf.cpp.o` which will count the actual code and data size. (although parts of it will be cut out at link time due to `-ffunction-sections`).

I'm having trouble getting a stable benchmark on my laptop, I might have to do a clean Linux boot with turbo boost and power management off. Windows doesn't keep a stable frequency.