@easyaspi314 2023-05-04
I think that there is a microoptimization issue, GCC x64 seems to get better performance with a different branch pattern.