@easyaspi314 2023-05-09
Drafting this for now, as I want to improve the intrinsics first before touching the table so I can get a better performance analysis.